refactoring & add test files
