+++ test/app/json/merge-requests/merge-close.json
@@ -1,0 +1,137 @@
+{
+    "object_kind": "merge_request",
+    "event_type": "merge_request",
+    "user": {
+      "name": "Jonas Michaelis",
+      "username": "jonmi",
+      "avatar_url": "https://assets.gitlab-static.net/uploads/-/system/user/avatar/2449793/avatar.png",
+      "email": "jonas.michaelis@outlook.com"
+    },
+    "project": {
+      "id": 18709877,
+      "name": "webhooks-test",
+      "description": "",
+      "web_url": "https://gitlab.com/jonmi/webhooks-test",
+      "avatar_url": null,
+      "git_ssh_url": "git@gitlab.com:jonmi/webhooks-test.git",
+      "git_http_url": "https://gitlab.com/jonmi/webhooks-test.git",
+      "namespace": "Jonas Michaelis",
+      "visibility_level": 0,
+      "path_with_namespace": "jonmi/webhooks-test",
+      "default_branch": "master",
+      "ci_config_path": null,
+      "homepage": "https://gitlab.com/jonmi/webhooks-test",
+      "url": "git@gitlab.com:jonmi/webhooks-test.git",
+      "ssh_url": "git@gitlab.com:jonmi/webhooks-test.git",
+      "http_url": "https://gitlab.com/jonmi/webhooks-test.git"
+    },
+    "object_attributes": {
+      "assignee_id": null,
+      "author_id": 2449793,
+      "created_at": "2020-05-12 05:42:06 UTC",
+      "description": "",
+      "head_pipeline_id": null,
+      "id": 58356482,
+      "iid": 1,
+      "last_edited_at": null,
+      "last_edited_by_id": null,
+      "merge_commit_sha": null,
+      "merge_error": null,
+      "merge_params": {
+        "force_remove_source_branch": "1"
+      },
+      "merge_status": "cannot_be_merged",
+      "merge_user_id": null,
+      "merge_when_pipeline_succeeds": false,
+      "milestone_id": null,
+      "source_branch": "dev",
+      "source_project_id": 18709877,
+      "state_id": 2,
+      "target_branch": "master",
+      "target_project_id": 18709877,
+      "time_estimate": 0,
+      "title": "Dev",
+      "updated_at": "2020-05-18 09:46:02 UTC",
+      "updated_by_id": null,
+      "url": "https://gitlab.com/jonmi/webhooks-test/-/merge_requests/1",
+      "source": {
+        "id": 18709877,
+        "name": "webhooks-test",
+        "description": "",
+        "web_url": "https://gitlab.com/jonmi/webhooks-test",
+        "avatar_url": null,
+        "git_ssh_url": "git@gitlab.com:jonmi/webhooks-test.git",
+        "git_http_url": "https://gitlab.com/jonmi/webhooks-test.git",
+        "namespace": "Jonas Michaelis",
+        "visibility_level": 0,
+        "path_with_namespace": "jonmi/webhooks-test",
+        "default_branch": "master",
+        "ci_config_path": null,
+        "homepage": "https://gitlab.com/jonmi/webhooks-test",
+        "url": "git@gitlab.com:jonmi/webhooks-test.git",
+        "ssh_url": "git@gitlab.com:jonmi/webhooks-test.git",
+        "http_url": "https://gitlab.com/jonmi/webhooks-test.git"
+      },
+      "target": {
+        "id": 18709877,
+        "name": "webhooks-test",
+        "description": "",
+        "web_url": "https://gitlab.com/jonmi/webhooks-test",
+        "avatar_url": null,
+        "git_ssh_url": "git@gitlab.com:jonmi/webhooks-test.git",
+        "git_http_url": "https://gitlab.com/jonmi/webhooks-test.git",
+        "namespace": "Jonas Michaelis",
+        "visibility_level": 0,
+        "path_with_namespace": "jonmi/webhooks-test",
+        "default_branch": "master",
+        "ci_config_path": null,
+        "homepage": "https://gitlab.com/jonmi/webhooks-test",
+        "url": "git@gitlab.com:jonmi/webhooks-test.git",
+        "ssh_url": "git@gitlab.com:jonmi/webhooks-test.git",
+        "http_url": "https://gitlab.com/jonmi/webhooks-test.git"
+      },
+      "last_commit": {
+        "id": "cd93f4f0af615190d33d968277e304eb31a1d6f3",
+        "message": "Initial commit",
+        "title": "Initial commit",
+        "timestamp": "2020-05-11T12:08:12+00:00",
+        "url": "https://gitlab.com/jonmi/webhooks-test/-/commit/cd93f4f0af615190d33d968277e304eb31a1d6f3",
+        "author": {
+          "name": "Jonas Michaelis",
+          "email": "jonas.michaelis@outlook.com"
+        }
+      },
+      "work_in_progress": false,
+      "total_time_spent": 0,
+      "human_total_time_spent": null,
+      "human_time_estimate": null,
+      "assignee_ids": [
+  
+      ],
+      "state": "closed",
+      "action": "close"
+    },
+    "labels": [
+  
+    ],
+    "changes": {
+      "state_id": {
+        "previous": 1,
+        "current": 2
+      },
+      "updated_at": {
+        "previous": "2020-05-12 05:42:06 UTC",
+        "current": "2020-05-18 09:46:02 UTC"
+      },
+      "total_time_spent": {
+        "previous": null,
+        "current": 0
+      }
+    },
+    "repository": {
+      "name": "webhooks-test",
+      "url": "git@gitlab.com:jonmi/webhooks-test.git",
+      "description": "",
+      "homepage": "https://gitlab.com/jonmi/webhooks-test"
+    }
+  }
